Set install_plots to use ImageMagick directly
Call out ImageMagick "convert" directly instead of using ps2any.  It
looks like the ghostscript device driver on CentOS6 is different from
CentOS5 and is producing unexpected results.  Applying the "-flatten"
operator in in convert in install_plots.pl appears to be a useful
temporary fix until we decide if ps2any should have the "-flatten"
flag added as well.